ethereum: fixed tokenByChainAddress() to use 32bit chain_id (#398)

+++ firmware/ethereum_tokens.c
@@ -722,7 +722,7 @@ const TokenType tokens[TOKENS_COUNT] = {
 
 const TokenType *UnknownToken = (const TokenType *)1;
 
-const TokenType *tokenByChainAddress(uint8_t chain_id, const uint8_t *address)
+const TokenType *tokenByChainAddress(uint32_t chain_id, const uint8_t *address)
 {
 	if (!address) return 0;
 	for (int i = 0; i < TOKENS_COUNT; i++) {
